Step jgit to 4.7 to fix change detection bug
Due to an incorrect evaluation of the path filtering logic
in jgit, paths that should have been returned was filtered
out. This result in 'no change' for modules that actually
have changes within the files that are not part of a
sub-module.

The 'xxx.txt' file below is a new file since last release.

The diff detector should detect a change for server-module
in Case A, but not in Case B:

Case A:
server-modules
server-modules/xxx.txt
server-modules/server-module-a

Case B:
server-modules
server-modules/server-module-a
server-modules/server-module-a/xxx.txt

Corrected and expanded the testcases for this detection logic
within the release plugin.

jgit 4.7 contains the correction for this.
jgit 4.7 also require java 8 minimum